// src/app/budget/model/consumption-data-interface.ts
export interface ConsumptionDataInterface {
  renovation: boolean;
  renovated_at: string;
  installed_power: string;
  equipment_power_percentage: string;
  gradation: string;
  gradation_hours: string;
  sub_total_consumption: string;
  new_sub_total_consumption: string;
  new_installed_power: string;
  new_equipment_power_percentage: string;
  new_gradation: string;
  new_gradation_hours: string;
  total_consumption: string;
  total: string;
}